pyproject.toml
[tool.poetry]
name = "dolang"
version = "0.0.11"
description = "Dolo Modeling Language"
readme = "README.md"
authors = ["Winant Pablo <pablo.winant@gmail.com>"]
license="BSD-2-Clause"

[tool.poetry.dependencies]
python = "^3.7"
numba = ">=0.50"
numpy = "^1.18.3"
sympy = "^1.5.1"
lark-parser = "^0.9.0"

[tool.poetry.dev-dependencies]
pyyaml = "^5.3.1"
pytest = "^5.2"
pylint = "^2.6.0"
mkdocs = "^1.1.2"
ipython = "^7.18.1"
ipykernel = "^5.3.4"

[build-system]
requires = ["poetry>=0.12"]
build-backend = "poetry.masonry.api"
